% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egor.R
\name{as.egor.egodata}
\alias{as.egor.egodata}
\title{Convert \code{\link{egodata}} Objects to \code{\link{egor}} Objects}
\usage{
as.egor.egodata(object, ...)
}
\arguments{
\item{object}{a \code{\link{egodata}} object}
}
\value{
An \code{\link{egor}} object.
}
\description{
Convert \code{\link{egodata}} Objects to \code{\link{egor}} Objects
}
\section{Miscellaneous Methods}{
 The following \dQuote{standard} methods have
}

\author{
Pavel N. Krivitsky
}
\keyword{manip}
\keyword{methods}
