% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egor.R
\name{as.egor.network}
\alias{as.egor.network}
\title{Construct an Egocentric View of a \code{\link{network}} Object}
\usage{
\method{as.egor}{network}(x, special.cols = c("na"), ...)
}
\arguments{
\item{x}{A \code{\link[network]{network}} object.}

\item{special.cols}{Vertex attributes that should not be copied to the
\code{egos} and \code{alters} tables. Defaults to attributes special to the
\code{\link[network]{network}} objects.}

\item{...}{Additional arguments, currently unused.}
}
\value{
An \code{\link{egor}} object.
}
\description{
Given a \code{\link[network]{network}} object, construct an
\code{\link{egor}} object representing a census of all the actors in the
network. Used mainly for testing.
}
\examples{

# See example(ergm.ego) and example(template_network).
}
\seealso{
\code{\link{template_network}}, which performs the inverse
operation (though drops the ties).
}
\author{
Pavel N. Krivitsky
}
\keyword{datagen}
\keyword{manip}
