% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EgoStat.R
\name{EgoStat-internal}
\alias{EgoStat-internal}
\alias{.allAlters}
\alias{.attrErr}
\alias{.unfactor}
\alias{.exsum}
\alias{.extabulate}
\alias{.matchNA}
\alias{.pasteNA}
\alias{.checkNA}
\title{Helper functions of EgoStats. Some may be documented and exported in the future.}
\usage{
.allAlters(egor)

.attrErr(term, attrname, req = c("one", "both"))

.unfactor(x)

.exsum(x)

.extabulate(x, nbins = max(1, bin, na.rm = TRUE))

.matchNA(x, table, nomatch = NA_integer_, incomparables = NULL)

.pasteNA(..., sep = " ", collapse = NULL)

.checkNA(x, ...)
}
\arguments{
\item{egor}{an \code{\link[=egor]{egor()}} object.}

\item{x}{a vector.}

\item{...}{Additional arguments to subroutines.}
}
\description{
Helper functions of EgoStats. Some may be documented and exported in the future.
}
\section{Functions}{
\itemize{
\item \code{.allAlters}: Return a \code{\link{tibble}} containing all alters.

\item \code{.attrErr}: Construct and throw an error message about ego or alter attributes being missing.

\item \code{.unfactor}: Convert a factor to its "ordinary" vector representation.

\item \code{.exsum}: As \code{\link[=sum]{sum()}}, but "extrapolating" the \code{NA}s.

\item \code{.extabulate}: As \code{\link[=tabulate]{tabulate()}}, but "extrapolating" the \code{NA}s.

\item \code{.matchNA}: As \code{\link[base:match]{base::match()}}, but \code{NA}s are passed through.

\item \code{.pasteNA}: As \code{\link[=paste]{paste()}}, but \code{NA}s are passed through rather than typeset.

\item \code{.checkNA}: Pass through the input if it does not contain mising values; otherwise stop with given error. Arguments after the first are passed through to \code{\link[=stop]{stop()}}.
}}

\keyword{internal}
