% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EgoStat.R, R/EgoStat.duration.R,
%   R/InitErgmTerm.R
\docType{methods}
\name{ergm.ego-terms}
\alias{ergm.ego-terms}
\alias{terms-ergm.ego}
\alias{ergm.ego.terms}
\alias{terms.ergm.ego}
\alias{ergm-terms}
\alias{ergm.terms}
\alias{terms-ergm}
\alias{terms.ergm}
\alias{EgoStat}
\alias{EgoStat.edges}
\alias{EgoStat.nodecov}
\alias{EgoStat.nodefactor}
\alias{EgoStat.nodematch}
\alias{EgoStat.nodemix}
\alias{EgoStat.absdiff}
\alias{EgoStat.degree}
\alias{EgoStat.degrange}
\alias{EgoStat.concurrent}
\alias{EgoStat.concurrentties}
\alias{EgoStat.degreepopularity}
\alias{EgoStat.transitiveties}
\alias{EgoStat.cyclicalties}
\alias{EgoStat.esp}
\alias{EgoStat.gwesp}
\alias{EgoStat.gwdegree}
\alias{EgoStat.mean.age}
\alias{InitErgmTerm.netsize.adj}
\title{\code{\link[ergm]{ergm}} Terms Implemented for
\code{\link{egor}}}
\usage{
EgoStat.edges(egor)

EgoStat.nodecov(egor, attrname)

EgoStat.nodefactor(egor, attrname, base = 1)

EgoStat.nodematch(egor, attrname, diff = FALSE, keep = NULL)

EgoStat.nodemix(egor, attrname, base = NULL)

EgoStat.absdiff(egor, attrname, pow = 1)

EgoStat.degree(egor, d, by = NULL, homophily = FALSE)

EgoStat.degrange(egor, from = NULL, to = Inf, by = NULL,
  homophily = FALSE)

EgoStat.concurrent(egor, by = NULL)

EgoStat.concurrentties(egor, by = NULL)

EgoStat.degreepopularity(egor)

EgoStat.transitiveties(egor, attrname = NULL)

EgoStat.cyclicalties(egor, attrname = NULL)

EgoStat.esp(egor, d)

EgoStat.gwesp(egor, decay = NULL, fixed = FALSE, cutoff = 30,
  alpha = NULL)

EgoStat.gwdegree(egor, decay = NULL, fixed = FALSE, cutoff = 30)

EgoStat.mean.age(egor, emptyval = 0)

InitErgmTerm.netsize.adj(nw, arglist, ...)
}
\arguments{
\item{egor, attrname, base, diff, keep, pow, d, by, homophily, from, to, decay, fixed, cutoff, alpha, emptyval, nw, arglist, ...}{arguments to terms. See \code{\link[ergm]{ergm-terms}}.}
}
\description{
This page describes the \code{\link[ergm]{ergm}} terms (and hence network
statistics) for which inference based on egocentrically sampled data is
implemented in \code{ergm.ego} package. Other packages may add their own
terms.

A custom term giving a linear combination of change statistics
useful for network size adjustment.
}
\details{
The current recommendation for any package implementing additional
egocentric calculator terms is to create a help file with a name or alias
\code{ergm.ego-terms}, so that \code{help("ergm.ego-terms")} will
list egocentric ERGM terms available from all loaded packages.
}
\section{Currently implemented egocentric statistics}{
 For each of these,
please see their respective package's \code{ergm-terms} help for meaning and
parameters. The simplest way to do this is usually via \code{? TERM}.

\describe{ \item{Special-purpose terms:}{ \describe{
\item{netsize.adj(edges=+1, mutual=0, transitiveties=0,
cyclicalties=0)}{A special-purpose term equivalent to a linear
combination of \code{\link[ergm]{edges}},
\code{\link[ergm]{mutual}}, \code{\link[ergm]{transitiveties}}, and
\code{\link[ergm]{cyclicalties}}, to house the network-size
adjustment offset. This term is added to the model automatically
and should not be used in the model formula directly.  } } }

\item{ergm:}{
\itemize{
\item \code{edges}
\item \code{nodecov}
\item \code{nodefactor}
\item \code{nodematch}
\item \code{nodemix}
\item \code{absdiff}
\item \code{degree}
\item \code{degrange}
\item \code{concurrent}
\item \code{concurrentties}
\item \code{degreepopularity}
\item \code{transitiveties}
\item \code{cyclicalties}
\item \code{esp}
\item \code{gwesp}
\item \code{gwdegree}
}
}

\item{tergm:}{
\itemize{
\item \code{mean.age}
}
}
}
}

\seealso{
\code{\link[ergm]{ergm-terms}}
}
\keyword{models}
