% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.ergmego.R
\name{predict.ergm.ego}
\alias{predict.ergm.ego}
\title{ERGM-based predicted tie probabilities for the pseudo-population network}
\usage{
\method{predict}{ergm.ego}(object, ...)
}
\arguments{
\item{object}{model fit as returned by \code{\link[=ergm.ego]{ergm.ego()}}}

\item{...}{other arguments passed to/from other methods}
}
\value{
See \code{\link[ergm:predict.formula]{ergm::predict.ergm()}}
}
\description{
ERGM-based predicted tie probabilities for the pseudo-population network
}
