% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egor.R
\name{sample}
\alias{sample}
\alias{sample.default}
\alias{sample.egor}
\title{Draw random egocentric subsamples}
\usage{
sample(x, size, replace = FALSE, prob = NULL, ...)

\method{sample}{default}(x, ...)

\method{sample}{egor}(x, size, replace = FALSE, prob = NULL, ...)
}
\arguments{
\item{x, size, replace, prob}{see \code{\link[base:sample]{base::sample()}}.}

\item{...}{extra arguments, currently unused.}
}
\description{
Implementations of the \code{\link[base:sample]{base::sample()}} function for \code{\link[egor:egor]{egor::egor()}} data.
}
\note{
A reimplementation of sample as a generic was necessary
because \code{\link[base:sample]{base::sample()}} is not a generic and cannot take
data-frame-alikes as arguments.
}
