% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egor.R
\name{as.egor.egodata}
\alias{as.egor.egodata}
\alias{egodata}
\alias{as_egor.egodata}
\title{Convert (deprecated) \code{\link{egodata}} Objects to
\code{\link{egor}} Objects}
\usage{
\method{as.egor}{egodata}(x, ...)

as_egor.egodata(x, ...)
}
\arguments{
\item{x}{a \code{\link{egodata}} object}

\item{...}{additional arguments, currently unused.}
}
\value{
An \code{\link{egor}} object.
}
\description{
Convert (deprecated) \code{\link{egodata}} Objects to
\code{\link{egor}} Objects
}
\author{
Pavel N. Krivitsky
}
\keyword{manip}
\keyword{methods}
